+++ src/com/magento/idea/magento2plugin/actions/generation/data/MagentoPluginMethodData.java
@@ -6,23 +6,17 @@ package com.magento.idea.magento2plugin.actions.generation.data;
 
 import com.jetbrains.php.lang.documentation.phpdoc.psi.PhpDocComment;
 import com.jetbrains.php.lang.psi.elements.Method;
-import com.jetbrains.php.lang.psi.elements.PhpClass;
-import org.jetbrains.annotations.NotNull;
 
 public class MagentoPluginMethodData {
-    private final PhpClass pluginClass;
     private final PhpDocComment docComment;
     private final Method method;
     private final Method targetMethod;
-    private final MagentoPluginMethodData.Type pluginType;
 
-    public MagentoPluginMethodData(Method targetMethod, PhpClass pluginClass, PhpDocComment docComment, Method method, @NotNull MagentoPluginMethodData.Type pluginType) {
+    public MagentoPluginMethodData(Method targetMethod, PhpDocComment docComment, Method method) {
         super();
-        this.pluginClass = pluginClass;
         this.docComment = docComment;
         this.method = method;
         this.targetMethod = targetMethod;
-        this.pluginType = pluginType;
     }
 
     public PhpDocComment getDocComment() {
@@ -33,24 +27,7 @@ public class MagentoPluginMethodData {
         return this.method;
     }
 
-    public PhpClass getPluginClass() {
-        return this.pluginClass;
-    }
-
     public Method getTargetMethod() {
         return this.targetMethod;
-    }
-
-    public Type getPluginType() {
-        return pluginType;
-    }
-
-    public static enum Type {
-        AROUND,
-        AFTER,
-        BEFORE;
-
-        private Type() {
-        }
     }
 }