+++ src/com/magento/idea/magento2plugin/util/magento/FileBasedIndexUtil.java
@@ -6,12 +6,16 @@ package com.magento.idea.magento2plugin.util.magento;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.psi.PsiFile;
+import com.intellij.psi.PsiManager;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.indexing.FileBasedIndex;
 import com.jetbrains.php.lang.PhpFileType;
+import com.magento.idea.magento2plugin.magento.packages.Package;
 import com.magento.idea.magento2plugin.stubs.indexes.ModuleNameIndex;
 import com.magento.idea.magento2plugin.util.RegExUtil;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.regex.Matcher;
@@ -41,6 +45,40 @@ public class FileBasedIndexUtil {
             viewVfs.addAll(getValues(moduleName, moduleVf, project));
         }
         return viewVfs;
+    }
+
+    public static PsiFile findModuleConfigFile(String virtualFieName, Package.Areas area, String moduleName, Project project)
+    {
+        Pattern pattern = Pattern.compile(RegExUtil.Magento.MODULE_NAME);
+        Matcher matcher = pattern.matcher(moduleName);
+        if (!matcher.find()) {
+            return null;
+        }
+
+        Collection<VirtualFile> moduleVfs =
+                FileBasedIndex.getInstance().getContainingFiles(ModuleNameIndex.KEY, moduleName,
+                        GlobalSearchScope.getScopeRestrictedByFileTypes(
+                                GlobalSearchScope.allScope(project),
+                                PhpFileType.INSTANCE
+                        )
+                );
+        if (moduleVfs.isEmpty()) {
+            return null;
+        }
+
+        VirtualFile moduleVf = moduleVfs.iterator().next();
+
+        String relativePath = File.separator.concat(Package.MODULE_BASE_AREA_DIR).concat(File.separator);
+        if (!area.equals(Package.Areas.base)) {
+            relativePath = relativePath.concat(area.toString()).concat(File.separator);
+        }
+        relativePath = relativePath.concat(virtualFieName);
+
+        VirtualFile configFile = moduleVf.getParent().findFileByRelativePath(relativePath);
+        if (configFile == null) {
+            return null;
+        }
+        return PsiManager.getInstance(project).findFile(configFile);
     }
 
     public static Collection<VirtualFile> findViewVfsByModuleVf(VirtualFile moduleVf, Project project)